+++ Geode/IOptions.cs
@@ -3,7 +3,7 @@
 	public interface IOptions
 	{
 		string Output { get; set; }
-		int PackFormat { get; set; }
+		string PackFormat { get; set; }
 		bool DumpIR { get; set; }
 		bool Debug { get; set; }
 	}