+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 * Explicit casting
 * Reordered some builtin type hierarchies
 * Refactored Geode into a separate project
+* Changed the default pack format to 88.0 (1.21.9)